Rename InputComponent to TextInput.
